# docs/swagger.yaml
basePath: /
definitions:
  handler.AssessmentRequest:
    properties:
      grip_position:
        type: integer
      left_value:
        type: number
      right_value:
        type: number
      type:
        type: integer
    type: object
...
  handler.CreateSessionRequest:
    properties:
      assessments:
        items:
          $ref: '#/definitions/handler.AssessmentRequest'
        type: array
      date:
        type: string
      duration:
        type: integer
      is_assessment:
        type: boolean
      name:
        type: string
      notes:
        type: string
      rep_datas:
        items:
          $ref: '#/definitions/handler.RepDataRequest'
        type: array
      repeater_reps:
        type: integer
      repeater_rest_time:
        type: integer
      repeater_set_rest:
        type: integer
      repeater_sets:
        type: integer
      repeater_split_hand:
        type: boolean
      repeater_work_time:
        type: integer
      session_type:
        type: integer
    type: object
...
  handler.RepDataRequest:
    properties:
      average_weight:
        type: number
      duration:
        type: integer
      grip_position:
        type: integer
      index:
        type: integer
      is_rest:
        type: boolean
      right_hand:
        type: boolean
      target_weight:
        type: number
    type: object
...
  handler.SessionResponse:
    properties:
      date:
        type: string
      duration:
        type: integer
      id:
        type: integer
      is_assessment:
        type: boolean
      name:
        type: string
      notes:
        type: string
      repeater_reps:
        type: integer
      repeater_rest_time:
        type: integer
      repeater_set_rest:
        type: integer
      repeater_sets:
        type: integer
      repeater_split_hand:
        type: boolean
      repeater_work_time:
        type: integer
      session_type:
        type: integer
      user_id:
        type: string
    type: object
...
  handler.UpdateSessionRequest:
    properties:
      duration:
        type: integer
      name:
        type: string
      notes:
        type: string
    type: object
...
host: api.crimpy.app
info:
  contact:
    email: support@crimpy.com
    name: API Support
  description: Backend API for Crimpy climbing training application
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://api.crimpy.app/terms
  title: Crimpy API
  version: "1.0"
paths:
...
  /api/sessions:
    get:
      consumes:
      - application/json
      description: Retrieve all training sessions for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of sessions
          schema:
            items:
              $ref: '#/definitions/handler.SessionResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all sessions
      tags:
      - Session
    post:
      consumes:
      - application/json
      description: Create a new training session for the authenticated user with optional
        rep data and assessments
      parameters:
      - description: Session details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Session created successfully
          schema:
            $ref: '#/definitions/handler.SessionResponse'
        "400":
          description: Invalid request or validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new session
      tags:
      - Session
  /api/sessions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a session. User must own the session unless they are an
        admin.
      parameters:
      - description: Session ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid session ID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Session not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a session
      tags:
      - Session
    get:
      consumes:
      - application/json
      description: Retrieve a specific session by ID with all related rep data and
        assessments. User must own the session unless they are an admin.
      parameters:
      - description: Session ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session details with rep_datas and assessments
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid session ID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Session not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a session by ID
      tags:
      - Session
    put:
      consumes:
      - application/json
      description: Update a session's name, notes, and duration. User must own the
        session unless they are an admin.
      parameters:
      - description: Session ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated session details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated session
          schema:
            $ref: '#/definitions/handler.SessionResponse'
        "400":
          description: Invalid request or session ID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Session not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a session
      tags:
      - Session
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"